MegaChip update
Added new MegaChip model, complete with sound.
Changed Display rendering to a model that queries the core for metrics so we can use 8 or 32bpp displays.
Added protection for overflow when loading a ROM that's too large for the memory.
Sound declicking now a separate proc.
Triple buffered the display via the PresentDisplay array
Added debug logging facility
Changed handling of Model types in the DFM; now builds menu on the fly